requirements.txt
Django==3.2.5
djangorestframework==3.12.4
django-cors-headers==3.7.0
python-dateutil==2.8.2